pyproject.toml
[project]
name = "shogi-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "select766", email = "select766@outlook.jp" }
]
requires-python = ">=3.12"
dependencies = [
    "cshogi>=0.9.2",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "cshogi>=0.9.2",
    "datasets>=3.6.0",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.3",
    "peft>=0.15.2",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "trl>=0.18.1",
]
